Make a dedicated .py file for fieldtrial_testing_config generation
